+++ app/src/main/java/com/mrerror/parachut/Models/PendingOrders/User.java
@@ -31,10 +31,10 @@ public class User implements Serializable
     private String address;
     @SerializedName("lat")
     @Expose
-    private Integer lat;
+    private Double lat;
     @SerializedName("long")
     @Expose
-    private Integer _long;
+    private Double _long;
     @SerializedName("delivery_time")
     @Expose
     private Object deliveryTime;
@@ -108,19 +108,19 @@ public class User implements Serializable
         this.address = address;
     }
 
-    public Integer getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
 
-    public Integer getLong() {
+    public Double getLong() {
         return _long;
     }
 
-    public void setLong(Integer _long) {
+    public void setLong(Double _long) {
         this._long = _long;
     }
